+++ tests/test_content_models.py
@@ -5,7 +5,7 @@ from django.db import models
 from cms.test_utils.testcases import CMSTestCase
 
 from djangocms_versioning import helpers
-from djangocms_versioning.helpers import replace_default_manager
+from djangocms_versioning.helpers import replace_manager
 from djangocms_versioning.managers import PublishedContentManagerMixin
 from djangocms_versioning.test_utils.people.models import Person
 
@@ -16,13 +16,13 @@ class ContentModelTestCase(CMSTestCase):
 
     def test_replace_default_manager(self):
         self.assertNotIn(PublishedContentManagerMixin, Person.objects.__class__.mro())
-        replace_default_manager(Person)
+        replace_manager(Person, "objects", PublishedContentManagerMixin)
         self.assertIn(PublishedContentManagerMixin, Person.objects.__class__.mro())
 
     def test_replace_default_manager_twice(self):
-        replace_default_manager(Person)
+        replace_manager(Person, "objects", PublishedContentManagerMixin)
 
-        with patch.object(helpers, "published_content_manager_factory") as mock:
-            replace_default_manager(Person)
+        with patch.object(helpers, "manager_factory") as mock:
+            replace_manager(Person, "objects", PublishedContentManagerMixin)
 
         mock.assert_not_called()